+++ simple/library/includes/classes/class-core.php
@@ -9,7 +9,7 @@
  * @subpackage  Core
  *
  * @since    1.0
- * @version  1.1
+ * @version  1.3
  */
 
 
@@ -20,7 +20,7 @@
  * Core class
  *
  * @since    1.0
- * @version  1.0
+ * @version  1.3
  *
  * Contents:
  *
@@ -49,19 +49,15 @@ final class {%= prefix_class %}_Theme_Framework {
 		 * Constructor
 		 *
 		 * @since    1.0
-		 * @version  1.0
+		 * @version  1.3
 		 */
 		private function __construct() {
 
 			// Processing
 
-				/**
-				 * Hooks
-				 */
+				// Hooks
 
-					/**
-					 * Actions
-					 */
+					// Actions
 
 						// Theme upgrade action
 
@@ -80,11 +76,7 @@ final class {%= prefix_class %}_Theme_Framework {
 
 							add_action( 'customize_save_after', array( $this, 'custom_styles_cache' ) );
 
-
-
-					/**
-					 * Filters
-					 */
+					// Filters
 
 						// Escape inline CSS
 
@@ -103,9 +95,13 @@ final class {%= prefix_class %}_Theme_Framework {
 
 							add_filter( 'the_content', array( $this, 'add_table_of_contents' ), 10 );
 
-						// Minify custom CSS
+						// Minify CSS
 
 							add_filter( 'wmhook_{%= prefix_hook %}_tf_custom_styles_output_cache', array( $this, 'minify_css' ), 10 );
+
+						// SSL ready URLs
+
+							add_filter( 'wmhook_{%= prefix_hook %}_tf_custom_styles_output', array( $this, 'fix_ssl_urls' ), 9999 );
 
 		} // /__construct
 
@@ -186,7 +182,7 @@ final class {%= prefix_class %}_Theme_Framework {
 		 * @link  http://blog.rrwd.nl/2014/11/21/html5-headings-in-wordpress-lets-fight/
 		 *
 		 * @since    1.0
-		 * @version  1.0.3
+		 * @version  1.3
 		 */
 		public static function get_the_logo() {
 
@@ -203,10 +199,17 @@ final class {%= prefix_class %}_Theme_Framework {
 
 				$output = '';
 
-				$document_title = ( 0 > version_compare( $GLOBALS['wp_version'], '4.4' ) ) ? ( wp_title( '|', false, 'right' ) ) : ( wp_get_document_title() ); // Compatible with WordPress 4.4
+				// @todo Remove `wp_title` with WordPress 4.6
+				$document_title = ( 0 > version_compare( $GLOBALS['wp_version'], '4.4' ) ) ? ( wp_title( '|', false, 'right' ) ) : ( wp_get_document_title() ); // Since WordPress 4.4
 
-				$jetpack_site_logo = get_option( 'site_logo', array() );
-				$jetpack_site_logo = ( function_exists( 'jetpack_get_site_logo' ) && isset( $jetpack_site_logo['id'] ) && $jetpack_site_logo['id'] ) ? ( absint( $jetpack_site_logo['id'] ) ) : ( false );
+				$custom_logo = get_theme_mod( 'custom_logo' ); // Since WordPress 4.5
+
+				// If we don't get WordPress 4.5+ custom logo, try Jetpack Site Logo
+
+					if ( empty( $custom_logo ) && function_exists( 'jetpack_get_site_logo' ) ) {
+						$custom_logo = get_option( 'site_logo', array() );
+						$custom_logo = ( isset( $custom_logo['id'] ) && $custom_logo['id'] ) ? ( absint( $custom_logo['id'] ) ) : ( false );
+					}
 
 				$blog_info = apply_filters( 'wmhook_{%= prefix_hook %}_tf_get_the_logo_blog_info', array(
 						'name'        => trim( get_bloginfo( 'name' ) ),
@@ -214,7 +217,7 @@ final class {%= prefix_class %}_Theme_Framework {
 					) );
 
 				$args = apply_filters( 'wmhook_{%= prefix_hook %}_tf_get_the_logo_args', array(
-						'logo_image' => ( function_exists( 'jetpack_get_site_logo' ) && $jetpack_site_logo ) ? ( array( $jetpack_site_logo ) ) : ( false ),
+						'logo_image' => ( ! empty( $custom_logo ) ) ? ( array( $custom_logo ) ) : ( false ),
 						'logo_type'  => 'text',
 						'title_att'  => ( $blog_info['description'] ) ? ( $blog_info['name'] . ' | ' . $blog_info['description'] ) : ( $blog_info['name'] ),
 						'url'        => home_url( '/' ),
@@ -225,9 +228,9 @@ final class {%= prefix_class %}_Theme_Framework {
 
 				// Logo image
 
-					if ( ! empty( $args['logo_image'] ) ) {
+					if ( ! empty( $args['logo_image'] ) && $args['logo_image'][0] ) {
 
-						$img_id = ( is_numeric( $args['logo_image'] ) ) ? ( absint( $args['logo_image'] ) ) : ( self::get_image_id_from_url( $args['logo_image'] ) );
+						$img_id = ( is_numeric( $args['logo_image'][0] ) ) ? ( absint( $args['logo_image'][0] ) ) : ( self::get_image_id_from_url( $args['logo_image'][0] ) );
 
 						if ( $img_id ) {
 
@@ -324,7 +327,7 @@ final class {%= prefix_class %}_Theme_Framework {
 		 * Appends the output at the top and bottom of post content.
 		 *
 		 * @since    1.0
-		 * @version  1.0.2
+		 * @version  1.3
 		 *
 		 * @param  string $content
 		 */
@@ -356,10 +359,10 @@ final class {%= prefix_class %}_Theme_Framework {
 				$title      = apply_filters( 'wmhook_{%= prefix_hook %}_tf_add_table_of_contents_title', '<h2 class="screen-reader-text">' . $title_text . '</h2>' );
 
 				$args = apply_filters( 'wmhook_{%= prefix_hook %}_tf_add_table_of_contents_args', array(
-						'disable_first' => true, //First part to have a title of the post (part title won't be parsed)?
-						'links'         => array(), //The output HTML links
-						'post_content'  => ( isset( $post->post_content ) ) ? ( $post->post_content ) : ( '' ), //Get the whole post content
-						'tag'           => 'h2', //HTML heading tag to parse as a post part title
+						'disable_first' => true, // First part to have a title of the post (part title won't be parsed)?
+						'links'         => array(), // The output HTML links
+						'post_content'  => ( isset( $post->post_content ) ) ? ( $post->post_content ) : ( '' ), // Get the whole post content
+						'tag'           => 'h2', // HTML heading tag to parse as a post part title
 					) );
 
 				// Post part counter
@@ -438,11 +441,12 @@ final class {%= prefix_class %}_Theme_Framework {
 		 * Get the post meta info
 		 *
 		 * hAtom microformats compatible. @link http://goo.gl/LHi4Dy
+		 * Supports WP ULike plugin. @link https://wordpress.org/plugins/wp-ulike/
 		 * Supports ZillaLikes plugin. @link http://www.themezilla.com/plugins/zillalikes/
 		 * Supports Post Views Count plugin. @link https://wordpress.org/plugins/baw-post-views-count/
 		 *
 		 * @since    1.0
-		 * @version  1.1
+		 * @version  1.3
 		 *
 		 * @param  array $args
 		 */
@@ -587,19 +591,32 @@ final class {%= prefix_class %}_Theme_Framework {
 							break;
 							case 'likes':
 
-								if (
-										apply_filters( 'wmhook_{%= prefix_hook %}_tf_get_the_post_meta_info_enable_' . $meta, true, $args )
-										&& function_exists( 'zilla_likes' )
-									) {
-									global $zilla_likes;
-									$helper = $zilla_likes->do_likes();
+								if ( apply_filters( 'wmhook_{%= prefix_hook %}_tf_get_the_post_meta_info_enable_' . $meta, true, $args ) ) {
 
-									$replacements = array(
-											'{attributes}'  => '',
-											'{class}'       => esc_attr( 'entry-likes entry-meta-element' ),
-											'{description}' => '',
-											'{content}'     => $helper,
-										);
+									if ( function_exists( 'wp_ulike' ) ) {
+									// WP ULike first
+
+										$replacements = array(
+												'{attributes}'  => '',
+												'{class}'       => esc_attr( 'entry-likes entry-meta-element' ),
+												'{description}' => '',
+												'{content}'     => wp_ulike( 'put' ),
+											);
+
+									} elseif ( function_exists( 'zilla_likes' ) ) {
+									// ZillaLikes after
+
+										global $zilla_likes;
+
+										$replacements = array(
+												'{attributes}'  => '',
+												'{class}'       => esc_attr( 'entry-likes entry-meta-element' ),
+												'{description}' => '',
+												'{content}'     => $zilla_likes->do_likes(),
+											);
+
+									}
+
 								}
 
 							break;
@@ -1148,7 +1165,7 @@ final class {%= prefix_class %}_Theme_Framework {
 		 * @uses  `wmhook_{%= prefix_hook %}_custom_styles` global hook
 		 *
 		 * @since    1.0
-		 * @version  1.1
+		 * @version  1.3
 		 *
 		 * @param  bool $set_cache  Determines whether the results should be cached or not.
 		 * @param  bool $return     Whether to return a value or just run the process.
@@ -1176,7 +1193,7 @@ final class {%= prefix_class %}_Theme_Framework {
 				$theme_options = (array) apply_filters( 'wmhook_{%= prefix_hook %}_theme_options', array() );
 				$alphas        = array_filter( (array) apply_filters( 'wmhook_{%= prefix_hook %}_tf_custom_styles_alphas', array() ) );
 
-				$replacements  = array_unique( array_filter( (array) get_transient( '{%= theme_slug %}_customizer_values' ) ) ); //There have to be values (defaults) set!
+				$replacements  = array_unique( array_filter( (array) get_transient( '{%= theme_slug %}_customizer_values' ) ) ); // There have to be values (defaults) set!
 
 				/**
 				 * Force caching during the first theme display when no cache set (default
@@ -1366,7 +1383,10 @@ final class {%= prefix_class %}_Theme_Framework {
 					}
 
 
+
 			// Output
+
+				$output = (string) apply_filters( 'wmhook_{%= prefix_hook %}_tf_custom_styles_output', $output );
 
 				if ( $output && $return ) {
 					return trim( (string) $output );
@@ -1421,18 +1441,27 @@ final class {%= prefix_class %}_Theme_Framework {
 	 */
 
 		/**
-		 * Get (parent) theme folder name
+		 * Fixing URLs in `is_ssl()` returns TRUE
 		 *
-		 * @since    1.0.15
-		 * @version  1.0.15
+		 * @since    1.3
+		 * @version  1.3
+		 *
+		 * @param  string $content
 		 */
-		public static function get_theme_slug() {
+		static public function fix_ssl_urls( $content ) {
+
+			// Processing
+
+				if ( is_ssl() ) {
+					$content = str_ireplace( 'http:', 'https:', $content );
+				}
+
 
 			// Output
 
-				return ( is_child_theme() ) ? ( wp_get_theme()->parent()->get_template() ) : ( null );
+				return $content;
 
-		} // /get_theme_slug
+		} // /fix_ssl_urls
 
 
 
@@ -1653,7 +1682,7 @@ final class {%= prefix_class %}_Theme_Framework {
 		 * Returns true if a blog has more than 1 category
 		 *
 		 * @since    1.0
-		 * @version  1.0
+		 * @version  1.3
 		 */
 		public static function is_categorized_blog() {
 
@@ -1675,7 +1704,7 @@ final class {%= prefix_class %}_Theme_Framework {
 						$all_cats = get_categories( array(
 								'fields'     => 'ids',
 								'hide_empty' => 1,
-								'number'     => 2, //we only need to know if there is more than one category
+								'number'     => 2, // We only need to know if there is more than one category
 							) );
 
 					// Count the number of categories that are attached to the posts